ControlledNumberTextField 作成
